+++ auxilliary_application/main.py
@@ -1,4 +1,4 @@
-from auxilliary_application.key_gen_app.gui import Gui
+from auxilliary_application.gui import Gui
 
 
 if __name__ == "__main__":
